refactor: LottoChecker 삭제 후 메서드들 LottoResultsType 으로 이동
